reworked player parrying
- fixed boss projectile tweaking from disposed bitmaps
- bouncy projectiles limit to 3 bounces before disposing

+++ drawedOut/attacks/CustomProjectile.cs
@@ -6,7 +6,7 @@ namespace drawedOut
         private static HashSet<Enemy> _toDamage = new HashSet<Enemy>();
         private int _heal; 
 
-        public PlayerUltProjectile(PointF origin, int width, int height, float velocity, double angle, 
+        public PlayerUltProjectile(PointF origin, int width, int height, int velocity, double angle, 
                 Entity parent, Bitmap sprite, int dmg, int accel, int maxSpeed, int heal, bool fromPlayer=true)
             : base(origin:origin, width:width, height:height, velocity:velocity, angle:angle, xDiff:1, yDiff:1,
                     parent:parent, sprite:sprite, isLethal:true, dmg:dmg, accel:accel, maxSpeed:maxSpeed)
